requirements.txt
django==4.2.25
djangorestframework==3.16.1
